mudando app.js para smart component
